s3_saver/pyproject.toml
[tool.poetry]
name = "s3-saver"
version = "0.1.0"
description = ""
authors = ["Dmitry Patin <multik.patin@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic-settings = "^2.2.1"
fastapi = "^0.111.0"
fastapi-limiter = "^0.1.6"
uvicorn = "^0.30.1"
gunicorn = "^21.2.0"
redis = "^5.0.6"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.11"
mypy = "^1.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.lint]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "UP", # pyupgrade
    "F", # pyflakes
    "I", # isort
    "C", # flake8-comprehensions
    "B", # flake8-bugbear
    "SIM", # flake8-simplify
]
ignore = [
    "B905",
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

fixable = ["ALL"]
unfixable = []

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".sh",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "migrations",
    "*/migrations/",
    "*/settings.py",
    "tasks",
    "tests",
]
per-file-ignores = { }

line-length = 80

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

target-version = "py311"

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 60

[tool.mypy]
python_version = "3.11"
plugins = [
    "pydantic.mypy",
    "sqlalchemy.ext.mypy.plugin",
]
follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
explicit_package_bases = true
no_implicit_reexport = true
ignore_missing_imports = true
exclude = [
    "docs",
    "migrations",
    "tests",
    "static"
]
